comments and structure
Implemented extensive coments for indicators
reworked project structure to match
